@@ -1,7 +1,7 @@
 <Project Sdk="Microsoft.NET.Sdk.Web">
 
   <PropertyGroup>
-    <TargetFramework>netcoreapp2.1</TargetFramework>
+    <TargetFramework>net5.0</TargetFramework>
   </PropertyGroup>
 
   <ItemGroup>
@@ -9,8 +9,43 @@
   </ItemGroup>
 
   <ItemGroup>
-    <PackageReference Include="Microsoft.AspNetCore.App" />
-    <PackageReference Include="Microsoft.VisualStudio.Web.CodeGeneration.Design" Version="2.1.1" />
+    <None Include="wwwroot\layui\layui\font\iconfont.svg" />
+    <None Include="wwwroot\layui\layui\font\iconfont.woff2" />
+    <None Include="wwwroot\layui\layui\layui.all.js" />
+    <None Include="wwwroot\layui\layui\layui.js" />
+    <None Include="wwwroot\layui\layui\lay\modules\carousel.js" />
+    <None Include="wwwroot\layui\layui\lay\modules\code.js" />
+    <None Include="wwwroot\layui\layui\lay\modules\colorpicker.js" />
+    <None Include="wwwroot\layui\layui\lay\modules\element.js" />
+    <None Include="wwwroot\layui\layui\lay\modules\flow.js" />
+    <None Include="wwwroot\layui\layui\lay\modules\form.js" />
+    <None Include="wwwroot\layui\layui\lay\modules\jquery.js" />
+    <None Include="wwwroot\layui\layui\lay\modules\laydate.js" />
+    <None Include="wwwroot\layui\layui\lay\modules\layedit.js" />
+    <None Include="wwwroot\layui\layui\lay\modules\layer.js" />
+    <None Include="wwwroot\layui\layui\lay\modules\laypage.js" />
+    <None Include="wwwroot\layui\layui\lay\modules\laytpl.js" />
+    <None Include="wwwroot\layui\layui\lay\modules\mobile.js" />
+    <None Include="wwwroot\layui\layui\lay\modules\rate.js" />
+    <None Include="wwwroot\layui\layui\lay\modules\slider.js" />
+    <None Include="wwwroot\layui\layui\lay\modules\table.js" />
+    <None Include="wwwroot\layui\layui\lay\modules\transfer.js" />
+    <None Include="wwwroot\layui\layui\lay\modules\tree.js" />
+    <None Include="wwwroot\layui\layui\lay\modules\upload.js" />
+    <None Include="wwwroot\layui\layui\lay\modules\util.js" />
+    <None Include="wwwroot\layui\文档与示例\仓库\Gitee 码云.url" />
+    <None Include="wwwroot\layui\文档与示例\仓库\GitHub.url" />
+    <None Include="wwwroot\layui\文档与示例\文档.url" />
+    <None Include="wwwroot\layui\文档与示例\示例.url" />
+    <None Include="wwwroot\layui\更新日志.url" />
+  </ItemGroup>
+
+  <ItemGroup>
+    <PackageReference Include="Microsoft.AspNetCore.Identity.EntityFrameworkCore" Version="5.0.1" />
+    <PackageReference Include="Microsoft.VisualStudio.Web.CodeGeneration.Design" Version="5.0.1" />
+    <PackageReference Include="Pomelo.EntityFrameworkCore.MySql" Version="5.0.0-alpha.2" />
+    <PackageReference Include="RestSharp" Version="106.11.8-alpha.0.12" />
+    <PackageReference Include="System.Drawing.Common" Version="5.0.0" />
   </ItemGroup>
 
 </Project>
