@@ -1,19 +1,14 @@
 ﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
 using Microsoft.AspNetCore.Builder;
 using Microsoft.AspNetCore.Hosting;
-using Microsoft.AspNetCore.Http;
-using Microsoft.AspNetCore.HttpsPolicy;
 using Microsoft.AspNetCore.Identity;
-using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
 using FlowerHouse.Models;
 using FlowerHouse.Models.Entity;
 using FlowerHouse.BgService;
+using Microsoft.Extensions.Hosting;
 
 namespace FlowerHouse
 {
@@ -40,9 +35,17 @@ namespace FlowerHouse
 
             .AddEntityFrameworkStores<FlowerHouseContext>().AddDefaultTokenProviders();
 
-            services.AddDbContext<FlowerHouseContext>(options => options.UseMySql("Data Source=localhost;Database=flowerhouse;User Id=root;Password=juno57;port=3306;SslMode=None;"));
+            services.AddDbContextPool<FlowerHouseContext>(options => options.UseMySql("Data Source=localhost;Database=flowerhouse;User Id=root;Password=sin3306;port=3306;SslMode=None;",
+                new MySqlServerVersion(new Version(8, 0, 21)),
+                mysqlOptions =>
+                    mysqlOptions.CharSetBehavior(Pomelo.EntityFrameworkCore.MySql.Infrastructure.CharSetBehavior.NeverAppend))
+                .EnableSensitiveDataLogging()
+                .EnableDetailedErrors()
+            );
+
 
             services.AddDistributedMemoryCache();
+
             services.AddSession();
 
             services.AddAuthorization(options =>
@@ -57,42 +60,55 @@ namespace FlowerHouse
 
             services.AddHostedService<OrderTimeOutService>();
 
-            services.AddMvc().SetCompatibilityVersion(CompatibilityVersion.Latest);
+            services.AddControllersWithViews();
         }
 
         // This method gets called by the runtime. Use this method to configure the HTTP request pipeline.
-        public void Configure(IApplicationBuilder app, IHostingEnvironment env)
+        public void Configure(IApplicationBuilder app, IWebHostEnvironment env)
         {
             if (env.IsDevelopment())
             {
                 app.UseDeveloperExceptionPage();
             }
-            //else
-            //{
-            //    app.UseExceptionHandler("/Home/Error");
-            //    app.UseHsts();
-            //}
+            else
+            {
+                app.UseExceptionHandler("/Home/Error");
+                app.UseHsts();
+            }
 
 
             app.UseStaticFiles();
 
+            app.UseRouting();
+
             app.UseAuthentication();
+
+            app.UseAuthorization();
 
             app.UseSession();
 
-            app.UseMvc(routes =>
-            {
-                routes.MapRoute(
-                        name: "areas",
-                        template: "{area:exists}/{controller=Users}/{action=Index}/{id?}"
-                  );
+            SampleData.InitializeBlogDatabaseAsync(app.ApplicationServices).Wait();
 
-                routes.MapRoute(
+
+            app.UseEndpoints(endpoints =>
+            {
+                endpoints.MapControllerRoute(
                     name: "default",
-                    template: "{controller=Home}/{action=Index}/{id?}");
+                    pattern: "{controller=Home}/{action=Index}/{id?}");
             });
 
-            SampleData.InitializeBlogDatabaseAsync(app.ApplicationServices).Wait();
+            //app.UseMvc(routes =>
+            //{
+            //    routes.MapRoute(
+            //            name: "areas",
+            //            template: "{area:exists}/{controller=Users}/{action=Index}/{id?}"
+            //      );
+
+            //    routes.MapRoute(
+            //        name: "default",
+            //        template: "{controller=Home}/{action=Index}/{id?}");
+            //});
+
         }
     }
 }
